Add shop rooms stacking
