+++ JamGame/Assets/Scripts/Level/Room/CoreModel.cs
@@ -96,5 +96,10 @@ namespace Level.Room
         {
             return InstantiateCoreModel(new TileConfig(uid, Vector2Int.zero, 0));
         }
+
+        public bool CompareUid(CoreModel coreModel)
+        {
+            return uid == coreModel.Uid;
+        }
     }
 }
