+++ JamGame/Assets/Scripts/Level/Shop/View.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 43fcec0c0f7671448b69371954cd312f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
