+++ JamGame/Assets/Scripts/Level/Shop/View/Employees.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 47f96fa9971af044fb20dc5de868b1ca
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
