+++ JamGame/Assets/Scripts/Level/Shop/View/Rooms.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 00c458ed82725014ab18419e3db0c35a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
